Fixing bug triggering unnecessary auth calls
getSession() was requesting updated session data when run client side on every page load while logged in (instead of pulling it from sessionStore).

This was secure, but triggered an unessary ajax request when visiting a page that implmented sessions while logged in, which impacted page load times.

Fixing the bug raised another issue which was that sessionStorage is isolated between tabs, so the example now uses localStorage instead of sessionStorage for non-secure data (the user profile info).

Changes were made to accomodate the impact this has so that the auth system still works without JavaScript on the client and without access to localStorage (such as in private browsing mode).