Removing unused dependacy in layout
Also updating layout to pass null session object if one is not defined for any reason (i.e. hasn’t been passed to it) which doesn’t happen in the example but is good practice and prevents an error occuring if someone were to make a new page that uses the Layout component but not extend it from the Page component (or set this.props.session themselves).