Client now verifies session on server on first run
This is a small update to have the client store the session in window.session.

The purpose of the update is to prevent the client from appearing to still be logged in, even though the session on the server may have been deleted.

When the client first tires to read the session it checks to see if this is the first time it has read it (i.e. if this is the initial client side load) and if so it fetches the session from the server to check it is not stale / expired.

Ideally the server would return something “expires” value in the JSON returned by a call to the /session route, and we’d just check against that value in getSession() to see if we are over due to revalidate but this is an interium improvement.

Another option is to configure a “max age” value in the Session class.

Either method would have the advantage of not needing to pollute the global space in window.*